Module google_language1_beta1::api

Structs

AnalyzeEntitiesRequest
The entity analysis request message.

AnalyzeEntitiesResponse
The entity analysis response message.

AnalyzeSentimentRequest
The sentiment analysis request message.

AnalyzeSentimentResponse
The sentiment analysis response message.

AnalyzeSyntaxRequest
The syntax analysis request message.

AnalyzeSyntaxResponse
The syntax analysis response message.

AnnotateTextRequest
The request message for the text annotation API, which can perform multiple analysis types (sentiment, entities, and syntax) in one call.

AnnotateTextResponse
The text annotations response message.

CloudNaturalLanguage
Central instance to access all CloudNaturalLanguage related resource activities

DependencyEdge
Represents dependency parse tree information for a token.

Document
################################################################ # Represents the input to API methods.

DocumentAnalyzeEntityCall
Finds named entities (currently proper names and common nouns) in the text along with entity types, salience, mentions for each entity, and other properties.

DocumentAnalyzeSentimentCall
Analyzes the sentiment of the provided text.

DocumentAnalyzeSyntaxCall
Analyzes the syntax of the text and provides sentence boundaries and tokenization along with part of speech tags, dependency trees, and other properties.

DocumentAnnotateTextCall
A convenience method that provides all the features that analyzeSentiment, analyzeEntities, and analyzeSyntax provide in one call.

DocumentMethods
A builder providing access to all methods supported on document resources. It is not used directly, but through the CloudNaturalLanguage hub.

Entity
Represents a phrase in the text that is a known entity, such as a person, an organization, or location. The API associates information, such as salience and mentions, with entities.

EntityMention
Represents a mention for an entity in the text. Currently, proper noun mentions are supported.

Features
All available features for sentiment, syntax, and semantic analysis. Setting each one to true will enable that specific analysis for the input.

PartOfSpeech
Represents part of speech information for a token.

Sentence
Represents a sentence in the input document.

Sentiment
Represents the feeling associated with the entire text or entities in the text.

TextSpan
Represents an output piece of text.

Token
Represents the smallest syntactic building block of the text.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.